rename and remove cruft
