--- packages/api/prisma/schema.prisma
@@ -1,58 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  // provider = "postgresql"
-  // url      = env("DATABASE_URL")
-  provider = "sqlite"
-  url      = "file:./dev.db"
-}
-
-model User {
-  id                  String       @id @default(uuid())
-  createdAt           DateTime     @default(now())
-  updatedAt           DateTime     @default(now())
-  name                String
-  email               String
-  role                Role         @relation(fields: [roleName], references: [name])
-  team                String
-  skills              UserSkills[]
-  sentInvitations     Invitation[] @relation("inviteSender")
-  receivedInvitations Invitation[] @relation("inviteReceiver")
-  roleName            String
-}
-
-model Skill {
-  name  String       @id
-  users UserSkills[]
-}
-
-model UserSkills {
-  user       User     @relation(fields: [userId], references: [id])
-  userId     String // relation scalar field (used in the `@relation` attribute above)
-  skill      Skill    @relation(fields: [skillName], references: [name])
-  skillName  String
-  assignedAt DateTime @default(now())
-
-  @@id([userId, skillName])
-}
-
-model Role {
-  name  String @id
-  users User[]
-}
-
-model Invitation {
-  id            String @id @default(uuid())
-  sendingUser   User   @relation("inviteSender", fields: [senderId], references: [id])
-  receivingUser User   @relation("inviteReceiver", fields: [receiverId], references: [id])
-  status        String
-  message       String
-  team          String
-  senderId      String
-  receiverId    String
-}
