--- packages/api/src/context.ts
@@ -1,11 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-export interface Context {
-  prisma: PrismaClient
-}
-
-const prisma = new PrismaClient();
-
-export const context: Context = {
-  prisma,
-};
