+++ packages/api/src/objects/invitations/invitation.model.ts
@@ -1,0 +1,41 @@
+import { Model, Modifiers } from 'objection'
+import User from '../users/user.model'
+
+export default class Invitation extends Model {
+    id!: string
+    team!: string
+    status!: string
+    message?: string
+
+    senderId: string
+    sendingUser: User
+    receiverId: string
+    receivingUser: User
+
+
+    // Table name is the only required property.
+    static tableName = 'invitations'
+
+
+    // This object defines the relations to other models. The relationMappings
+    // property can be a thunk to prevent circular dependencies.
+    static relationMappings = () => ({
+        sendingUser: {
+            relation: Model.BelongsToOneRelation,
+            modelClass: User,
+            join: {
+                from: 'invitations.senderId',
+                to: 'roles.name',
+            },
+        },
+
+        receiverUser: {
+            relation: Model.BelongsToOneRelation,
+            modelClass: User,
+            join: {
+                from: 'invitations.receiverId',
+                to: 'roles.name',
+            },
+        },
+    })
+}
