+++ packages/api/src/objects/skills/skill.schema.ts
